Reinforcement 2
Encapsulating the learning process to allow both enemy and player to learn at the same time. This is building towards individual marbles having a Q table returning optimum strategies which the overall AI can then determine which to choose. This could also be done using one matrix which could enhance the performance.